+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - [#31](https://github.com/Open-Telecoms-Data/lib-cove-ofds/pull/31) - `jsontogeojson`: Dereference funders
 - Python Validate:
   - add more information to errors
+  - Add unique ID checks
 
 ## [0.4.0] - 2022-11-09